Use Cython if available in env
This removes the requirement to have "--cython" as an arg when doing "python setup.py install". If Cython is available in the current env, it will always try to build pybloomfilter.pyx. The "--cython" flag will now try to force Cython build, otherwise display an error message and exit if the module is not available.